refactor: seperate model from one file (#224)
* refactor: breakout model from one file

Model file was getting too big so broke them into separate files
To avoid circular dependency had to refactor Team.espys_total to hybrid property and moved get_teams_captain to Team model
